added Power Management PCB to repository
DO NOT CHANGE ANYTHING on Altium Project if you don't know what you are
doing. I have putten now more than 20h of work into the Power Management
PCB project to get it right this time.

I have drawn the first parts an created the structure of the project.
Everything is now on the best way to get the max. out of what I have in
mind.